==> Lab5/1_train_test_split.ipynb <==
import pandas as pd
import numpy as np
import random

with open(r"C:\Users\ashis\OneDrive\Desktop\NLP\LAB1\sentence_tokenized.txt", "r", encoding="utf-8") as f:
    sentences = [line.strip() for line in f if line.strip()]  

# Wrap in DataFrame for convenience
df = pd.DataFrame(sentences, columns=["sentence"])

# Show 10 random sentencesk
print(df.sample(10),"\n\n")
print("Total sentences:", len(df))

np.random.seed(42)

indices=np.random.permutation(len(df))

val_idx =indices[:1000]
test_idx=indices[1000:2000]
train_idx =indices[2000:]

val=df.iloc[val_idx]
test=df.iloc[test_idx]
train=df.iloc[train_idx]

print(len(train))
print(len(val))
print(len(test))

train.to_csv("train_sentences.csv", index=False, encoding="utf-8")
val.to_csv("val_sentences.csv", index=False, encoding="utf-8")
test.to_csv("test_sentences.csv", index=False, encoding="utf-8")

print("Data splits saved successfully.")
